% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bus_arrivals.R
\docType{data}
\name{bus_arrivals}
\alias{bus_arrivals}
\title{Bus arrival timings}
\format{A data frame with 103 rows and 30 variables}
\source{
Called from the Bus Timings API on 11 December 2018 at 12.06pm
}
\usage{
bus_arrivals
}
\description{
This is a small dataset containing bus arrival timings for 11 randomly selected bus stops in Singapore.
}
\keyword{datasets}
