% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bus_ridership.R
\docType{data}
\name{bus_ridership_cleaned}
\alias{bus_ridership_cleaned}
\title{Bus ridership}
\format{A data frame with 4734 rows and 6 variables}
\source{
Downloaded from the Passenger Volume by Bus Stop API, and joined with the Bus Stops API
}
\usage{
bus_ridership_cleaned
}
\description{
This is a dataset containing the total number of taps in and out for each bus stop in Singapore.
}
\keyword{datasets}
