% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bus_routes.R
\docType{data}
\name{bus_routes}
\alias{bus_routes}
\title{Bus routes}
\format{A data frame with 26080 rows and 12 variables}
\source{
Called from the Bus Routes API on 11 December 2018 at 11.20pm
}
\usage{
bus_routes
}
\description{
This is a dataset containing all the information for all bus routes in Singapore.
}
\keyword{datasets}
