% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bus_stops.R
\docType{data}
\name{bus_stops}
\alias{bus_stops}
\title{Bus Stops}
\format{A data frame with 5006 rows and 5 variables}
\source{
Called from the Bus Stops API on 12 December 2018 at 6.10am
}
\usage{
bus_stops
}
\description{
This is a dataset containing information on all the bus stops in Singapore.
}
\keyword{datasets}
