% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_carpark_avail.R
\docType{data}
\name{carpark_avail}
\alias{carpark_avail}
\title{Carpark Availability}
\format{A data frame with 2020 rows and 9 variables}
\source{
Called from the Bus Timings API on 12 December 2018 at 9.10am
}
\usage{
carpark_avail
}
\description{
This is a dataset containing information about all HDB, URA, and LTA carparks in Singapore and the number of available lots in each carpark.
}
\keyword{datasets}
