% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreSGMap.R
\name{exploreSGMap}
\alias{exploreSGMap}
\title{Explore data on an interactive map of Singapore}
\usage{
exploreSGMap(dataframe, popup = NA, cluster = FALSE, colour = "red",
  size = 5, alpha = 0.5)
}
\arguments{
\item{dataframe}{The dataframe containing the latitude/longitude data}

\item{popup}{Variable to be displayed when a marker is clicked on - default is NA, needs to be a string of the variable name}

\item{cluster}{Whether to cluster the markers or not - default is FALSE}

\item{colour}{Colour of the markers - default is 'red'}

\item{size}{Size of the markers - default is 5}

\item{alpha}{Opacity of the markers - default is 0.5}
}
\value{
An interactive map of Singapore with the coordinates plotted as circular markers on the map
}
\description{
A simple Leaflet wrapper which enables users to plot the data retrieved from the LTA APIs on an interactive map of Singapore. Some basic options are provided for ease of use. Note that this function only works automatically with outputs from getTaxiAvail, getCarparkAvail, getTrafficIncidents, getVMS, and getTrafficImages from this package. For all others, please obtain data for the coordinates and rename the respective columns to 'Latitude' and 'Longitude'.
}
\examples{
\donttest{
mydata <- getTrafficImages(mykey)
exploreSGMap(mydata, cluster = FALSE, colour = 'black', size = 7, alpha = 0.7)
}
}
