% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBusPassengerVol_od.R
\name{getBusPassengerVol_od}
\alias{getBusPassengerVol_od}
\title{Passenger Volume by Origin/Destination Bus Stops}
\usage{
getBusPassengerVol_od(date = Sys.Date() - 60, api_key)
}
\arguments{
\item{date}{Date to return the dataset for. The default is two months prior. Needs to be in the Date format.}

\item{api_key}{API key for LTA's Datamall}
}
\value{
A link from which the user can download the CSV file.
}
\description{
Returns a link from LTA that enables the user to download a CSV file containing a month's worth of data on passenger volume by origin/destination bus stops.
}
\examples{
getBusPassengerVol_od(Sys.Date()-60, Sys.getenv('LTA_DATAMALL_KEY'))
}
