% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBusRoutes.R
\name{getBusRoutes}
\alias{getBusRoutes}
\title{Bus Routes}
\usage{
getBusRoutes(api_key)
}
\arguments{
\item{api_key}{API key for LTA's Datamall}
}
\value{
A dataframe containing all the outputs from the Bus Routes API, such as the operator, stop sequence, bus stop code, and distance travelled.
}
\description{
Returns information on all bus routes in Singapore that are currently in operation.
}
\examples{
\donttest{
getBusRoutes(mykey)
}
}
