% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBusStops.R
\name{getBusStops}
\alias{getBusStops}
\title{Bus Stops}
\usage{
getBusStops(api_key)
}
\arguments{
\item{api_key}{API key for LTA's Datamall}
}
\value{
A dataframe containing all the outputs from the Bus Stops API, such as the bus stop code, road name, description, and coordinates.
}
\description{
Returns information on all bus stops in Singapore that are currently in operation.
}
\examples{
\donttest{
getBusStops(mykey)
}
}
