% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBusTimings.R
\name{getBusTimings}
\alias{getBusTimings}
\title{Bus Timings}
\usage{
getBusTimings(codes = c("07351", "09047"), api_key)
}
\arguments{
\item{codes}{A character vector of bus stop codes}

\item{api_key}{API key for LTA's Datamall}
}
\value{
A dataframe containing all the outputs from the Bus Timings API, including the service number, estimated arrival time, load, and type for the next three buses.
}
\description{
Returns real-time estimated bus arrival timings for the specified bus stops in Singapore. Note that this function may not return anything between 1.00am and 5.00am SG time (GMT+8)
}
\examples{
getBusTimings(c('11111', '22222', '33333'), Sys.getenv('LTA_DATAMALL_KEY'))
}
