% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getERPRates.R
\name{getERPRates}
\alias{getERPRates}
\title{ERP Rates}
\usage{
getERPRates(api_key)
}
\arguments{
\item{api_key}{API key for LTA's Datamall}
}
\value{
A dataframe containing ERP rates for each vehicle type, ERP rate timing, and zone.
}
\description{
Provides the ERP rates for all vehicle types across all timings and all zones.
}
\examples{
\donttest{
getERPRates(mykey)
}
}
