% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstTravelTime.R
\name{getEstTravelTime}
\alias{getEstTravelTime}
\title{Estimated Travel Times}
\usage{
getEstTravelTime(api_key)
}
\arguments{
\item{api_key}{API key for LTA's Datamall}
}
\value{
A dataframe containing each carpark's information and number of available lots
}
\description{
Provides key details and available lots for HDB, LTA, and URA carparks.
}
\examples{
\donttest{
getCarparkAvail(Sys.getenv('LTA_DATAMALL_KEY'))
}
}
