% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFaultyTrafficLights.R
\name{getFaultyTrafficLights}
\alias{getFaultyTrafficLights}
\title{Faulty Traffic Lights}
\usage{
getFaultyTrafficLights(api_key)
}
\arguments{
\item{api_key}{API key for LTA's Datamall}
}
\value{
A dataframe containing the alerts for each traffic light fault, such as the type of failure and start date-time
}
\description{
Returns alerts of traffic lights that are either currently faulty or are undergoing scheduled maintenance.
}
\examples{
\donttest{
getFaultyTrafficLights(mykey)
}
}
