% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRoadOpenings.R
\name{getRoadOpenings}
\alias{getRoadOpenings}
\title{Road Openings}
\usage{
getRoadOpenings(api_key)
}
\arguments{
\item{api_key}{API key for LTA's Datamall}
}
\value{
A dataframe containing the details for each planned road opening, such as the start date, end date, and new road name
}
\description{
Returns all planned road openings.
}
\examples{
\donttest{
getRoadOpenings(Sys.getenv('LTA_DATAMALL_KEY'))
}
}
