% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRoadWorks.R
\name{getRoadWorks}
\alias{getRoadWorks}
\title{Road Works}
\usage{
getRoadWorks(api_key)
}
\arguments{
\item{api_key}{API key for LTA's Datamall}
}
\value{
A dataframe containing the details for each current or planned road work, such as the start date, end date, and road name
}
\description{
Returns all road works currently being carried out or scheduled to be carried out.
}
\examples{
getRoadWorks(Sys.getenv('LTA_DATAMALL_KEY'))
}
