% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxiAvail.R
\name{getTaxiAvail}
\alias{getTaxiAvail}
\title{Taxi Availability}
\usage{
getTaxiAvail(api_key)
}
\arguments{
\item{api_key}{API key for LTA's Datamall}
}
\value{
A dataframe containing the latitude and longitude information for all the taxis available for hire.
}
\description{
Returns the coordinates for all taxis that are currently available for hire. Does not include 'hired' or 'busy' taxis.
}
\examples{
\donttest{
getTaxiAvail(mykey)
}
}
