% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrafficIncidents.R
\name{getTrafficIncidents}
\alias{getTrafficIncidents}
\title{Traffic Incidents}
\usage{
getTrafficIncidents(api_key)
}
\arguments{
\item{api_key}{API key for LTA's Datamall}
}
\value{
A dataframe containing each traffic incident, its type, coordinates, and description message
}
\description{
Returns incidents currently happening on the roads, such as accidents, vehicle breakdowns etc.
}
\examples{
\donttest{
getTrafficIncidents(Sys.getenv('LTA_DATAMALL_KEY'))
}
}
