% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrafficSpeedBands.R
\name{getTrafficSpeedBands}
\alias{getTrafficSpeedBands}
\title{Traffic Speed Bands}
\usage{
getTrafficSpeedBands(api_key)
}
\arguments{
\item{api_key}{API key for LTA's Datamall}
}
\value{
A dataframe containing each stretch of road, speed bands, and location data
}
\description{
Returns current traffic speeds on expressways and major roads, expressed in speed bands
}
\examples{
\donttest{
getTrafficSpeedBands(mykey)
}
}
