% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrainAlerts.R
\name{getTrainAlerts}
\alias{getTrainAlerts}
\title{Train Service Alerts}
\usage{
getTrainAlerts(api_key)
}
\arguments{
\item{api_key}{API key for LTA's Datamall}
}
\value{
If there is a train service disruption, it will return a dataframe containing the alerts and their respective details.
}
\description{
Returns alerts on train service unavailability during scheduled operating hours.
}
\examples{
\donttest{
getTrainAlerts(mykey)
}
}
