% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrainPassengerVol.R
\name{getTrainPassengerVol}
\alias{getTrainPassengerVol}
\title{Passenger Volume by Train Station}
\usage{
getTrainPassengerVol(date = Sys.Date() - 60, api_key)
}
\arguments{
\item{date}{Date to return the dataset for. The default is two months prior. Needs to be in the Date format.}

\item{api_key}{API key for LTA's Datamall}
}
\value{
A link from which the user can download the CSV file.
}
\description{
Returns a link from LTA that enables the user to download a CSV file containing a month's worth of data on passenger volume for each MRT station.
}
\examples{
\donttest{
getTrainPassengerVol(Sys.Date()-60, Sys.getenv('LTA_DATAMALL_KEY'))
}
}
