% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sg_map.R
\docType{data}
\name{sg_map}
\alias{sg_map}
\title{Singapore Map}
\format{A ggmap object}
\source{
Called from the Google Maps API on 12 December 2018 at 7.30am
}
\usage{
sg_map
}
\description{
This is a simple street map of Singapore, called from Google Maps API using the ggmap package.
}
\keyword{datasets}
