% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_taxi_avail.R
\docType{data}
\name{taxi_avail}
\alias{taxi_avail}
\title{Taxi Availability}
\format{A data frame with 4823 rows and 2 variables}
\source{
Called from the Taxi Availability API on 11 December 2018 at 11.10pm
}
\usage{
taxi_avail
}
\description{
This is a dataset containing the coordinates of all taxis available for hire in Singapore.
}
\keyword{datasets}
