% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSGMap.R
\name{plotSGMap}
\alias{plotSGMap}
\title{Plot data on a map of Singapore}
\usage{
plotSGMap(dataframe, popup = NA, cluster = FALSE, color = "red")
}
\arguments{
\item{dataframe}{The dataframe containing the latitude/longitude data}

\item{popup}{Name of the variable in the dataframe to be displayed when a marker is clicked on - default is NA}

\item{cluster}{Whether to cluster the markers or not - default is FALSE}

\item{color}{Color of the markers - default is 'red'}
}
\value{
A dataframe containing each carpark's information and number of available lots
}
\description{
A simple Leaflet wrapper which enables users to plot the data retrieved from the LTA APIs on an interactive map of Singapore. Some simple options are provided for ease of use. Note that this function only works automatically with outputs from getTaxiAvail, getCarparkAvail, getTrafficIncidents, getVMS, and getTrafficImages from this package. For all others, please rename the latitude and longitude columns to 'Latitude' and 'Longitude' respectively.
}
\examples{
\donttest{
mydata <- getTrafficImages(Sys.getenv('LTA_DATAMALL_KEY'))
plotSGMap(mydata, cluster = FALSE, color = 'black')
}
}
